+++ css/main.css
@@ -1,0 +1,20 @@
+header, footer {
+    background: radial-gradient(circle, orange, orange, dimgrey, black);
+    padding: 1em;
+    text-align: center;
+}
+
+footer a {
+    text-decoration: none;
+    color: black;
+}
+
+footer a:hover {
+    color: white;
+}
+
+.attraction-container {
+    background: radial-gradient(circle, dimgrey, black) fixed;
+    color: orange;
+    padding: 1em;
+}
